+++ oda_reader/download/download_tools.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 from oda_reader.common import api_response_to_df, logger
-from oda_reader.query_builder import QueryBuilder
+from oda_reader.download.query_builder import QueryBuilder
 from oda_reader.schemas.dac1_translation import convert_dac1_to_dotstat_codes
 from oda_reader.schemas.dac2_translation import convert_dac2a_to_dotstat_codes
 from oda_reader.schemas.schema_tools import (